+++ spec/lib/aixm/feature/navigational_aid/tacan_spec.rb
@@ -1,6 +1,21 @@
 require_relative '../../../../spec_helper'
 
 describe AIXM::Feature::NavigationalAid::TACAN do
+  CHANNELS = {
+    '1X' => AIXM.f(134.4, :mhz),
+    '12Y' => AIXM.f(135.55, :mhz),
+    '16Y' => AIXM.f(135.95, :mhz),
+    '17X' => AIXM.f(108, :mhz),
+    '30X' => AIXM.f(109.3, :mhz),
+    '59Y' => AIXM.f(112.25, :mhz),
+    '60X' => AIXM.f(133.3, :mhz),
+    '64Y' => AIXM.f(133.75, :mhz),
+    '69Y' => AIXM.f(134.25, :mhz),
+    '70X' => AIXM.f(112.30, :mhz),
+    '100X' => AIXM.f(115.3, :mhz),
+    '126Y' => AIXM.f(117.95, :mhz)
+  }
+
   subject do
     AIXM::Factory.tacan
   end
@@ -11,20 +26,30 @@ describe AIXM::Feature::NavigationalAid::TACAN do
     end
   end
 
+  describe :ghost_f= do
+    it "must set the corresponding channel" do
+      CHANNELS.each do |channel, f|
+        _(subject.tap { _1.ghost_f = f }.channel).must_equal channel
+        _(subject.tap { _1.ghost_f = f }.ghost_f).must_equal f
+      end
+    end
+
+    it "can be used to initialize a Tacan" do
+      same_dme = AIXM.tacan(
+        organisation: AIXM::Factory.organisation,
+        id: 'MMM',
+        xy: AIXM.xy(lat: %q(47°51'33"N), long: %q(007°33'36"E)),
+        ghost_f: AIXM.f(114.8, :mhz)
+      )
+      _(same_dme.channel).must_equal '95X'
+    end
+  end
+
   describe :ghost_f do
     it "must be derived from the channel" do
-      _(subject.tap { _1.channel = '1X' }.ghost_f.freq).must_equal 134.4
-      _(subject.tap { _1.channel = '12Y' }.ghost_f.freq).must_equal 135.55
-      _(subject.tap { _1.channel = '16Y' }.ghost_f.freq).must_equal 135.95
-      _(subject.tap { _1.channel = '17X' }.ghost_f.freq).must_equal 108
-      _(subject.tap { _1.channel = '30X' }.ghost_f.freq).must_equal 109.3
-      _(subject.tap { _1.channel = '59Y' }.ghost_f.freq).must_equal 112.25
-      _(subject.tap { _1.channel = '60X' }.ghost_f.freq).must_equal 133.3
-      _(subject.tap { _1.channel = '64Y' }.ghost_f.freq).must_equal 133.75
-      _(subject.tap { _1.channel = '69Y' }.ghost_f.freq).must_equal 134.25
-      _(subject.tap { _1.channel = '70X' }.ghost_f.freq).must_equal 112.30
-      _(subject.tap { _1.channel = '100X' }.ghost_f.freq).must_equal 115.3
-      _(subject.tap { _1.channel = '126Y' }.ghost_f.freq).must_equal 117.95
+      CHANNELS.each do |channel, f|
+        _(subject.tap { _1.channel = channel }.ghost_f).must_equal f
+      end
     end
   end